fix unused vars
When adjusting the linting rule for unused variables (to allow variables starting with _) I've noticed that some parts of the project weren't conforming to the linter. This attempts to fix it. I'm worried that other linting rules weren't active as well, due to us overriding .eslintrc, but I'm not sure we can pull off an eslint config that can lint the entire project, react stuff and all.