refactor ranges (and some other stuff in typesystem)
Ranges used to be just another property associated to a type. So a range of numbers was just a number with rangeness = true.

This changes it so that ranges are just another type of container, with a .rangeOf property indicating what it's made of.

Also changes which side is expected in cases like 'a + b' where the types are different. If 'a' is T1 and 'b' is T2, now we'll see 'Expected T1' instead of 'Expected T2'