Split app into pages
To have any decent kind of state management, we should organize the project in a way that allows us to.

In the future, we know 100% that there's only one active pad at a time. By splitting it into its own page, we can take advantage of this later.

This also removes some useRouter calls, which can cause us to be coupled to nextJS, a very costly mistake.

A Frame component was created, which is merely the layout of the page. It's better than hard-coding layout in pages/_app, because it's opt-in. For example the login page doesn't use the frame, so it can be a clean whole-page interface with no menus. A landing page would behave similarly.