Speed up runtime tests
Due to the usage of JSDOM, the tests got really slow -- jest requires JSDOM in each test file. If we do need JSDOM we can add a comment on top of the file that tells jest we are opting in to JSDOM on that file.

Also, there were some arbitrarily large timeouts, which I replaced with the npmjs.com/wait-for-expect library. It runs repeatedly until no jest-generated exceptions come out. After a timeout of 4.5 seconds, it borks with the appropriate exception. I could not always do this because some tests were testing a negative, and the timeout was essential.