feat(editor): disable the ability to format text outside of paragraph elements (#355)
* feat(editor): disable the ability to format text inside model blocks

This commit brings the following changes:

- Implements a custom onKeyDown event handler for the editor marks
(bold, italic and underline) that checks for the parent node's type
before deciding to format the text or not.
- Removes unused marks like the strikethrough.
- Modifies the marks' plugins to use the custom onKeyDown implementation.
- Installs the package @types/proper-lockfile since its absence caused the
type check command to fail.

On branch ahmed/modelblock-formatting
Changes to be committed:
	modified:   libs/editor/src/configuration/plugins.ts
	new file:   libs/editor/src/utils/getToggleMarkOnKeyDown.ts
	modified:   package-lock.json
	modified:   package.json

* refactor(editor): change from a blacklist to a whitelist

I have changed the checking for the block's type to be a whitelist,
this is for a couple of reasons:

- We don't want the user to be able to make the title or subtitle...etc
underlined, or italic, because this disturbs our design language. This
lead into the realisation that we only want users to change the text formatting
only when the block is a normal paragraph element. So that's why I added a check
that will only let the user format the text if the block they selected is a
paragraph element.

Changes to be committed:
	modified:   libs/editor/src/utils/getToggleMarkOnKeyDown.ts

* refactor(editor): isolate the marks into their own file

This commit brings the following changes:

- Groups all of the modified marks implementations into a single file
called createMarksPlugin that is used in the main plugins array. This is
better in terms of code organisation and also for future purposes when
we need to include more marks.

Changes to be committed:
	modified:   libs/editor/src/configuration/plugins.ts
	new file:   libs/editor/src/plugins/Marks/createMarksPlugin.ts

* feat(editor): disable autoformatting outside of paragraph elements

This commit brings the following changes:

- Disables the autoformatting outside of paragraph elements, this
prevents the user from doing something like _hello_ in a modelblock
or the title and for it to turn into underlined text.

- Removes unused autoformatting options, like for lists and todo items.
Changes to be committed:
	modified:   libs/editor/src/configuration/autoFormatOptions.ts

* fix(editor): create inline code plugin

This commit brings the following changes:

- Adds back the missing inline code mark in the editor with the new
implementation for the onKeyDown event.

Changes to be committed:
	modified:   libs/editor/src/plugins/Marks/createMarksPlugin.ts

* refactor(editor): naming