feat(editor): design system paragraph incl. placeholder
This sounds like very little, but actually this includes some
restructuring that establishes a basis for how we will structure
mostly Slate-independent UI components versus Slate specifics in the
editor.

Probably contains more random cleanups than it should have,
but it's still a very manageable size and most of that can be skimmed
through very quickly.

This also changes the placeholder to displayed on any empty paragraph,
focused or not (although it would be quite easy to implement the old
behavior), which I believe was a requirement I've seen somewhere.
This does currently mean if you have a lot of empty paragraphs you'll
see a lot of placeholders, which may not be desired, but I'm not sure
about exact requirements at the moment and we can address that later.