directives, part 1 (#535)
* directives, part 1

We've been splitting our evaluation and typechecking into two different folders src/infer and src/interpreter. When we create some special operations like 'as', which need access to the AST to return a value, we end up splitting them into two places.

For comparison, functions such as 'if' are declared in a much easier way because they're considered builtins. And they don't make the AST types more complicated.

To bridge this gap, I created directives, and made 'as' into the first directive.

This PR represents creating the directives system and using it for the first time. But it's not all! We will also need to change the AST slightly, such that directives no longer have their own dedicated AST node. To do so, there will be a 'directive' node to encompass any and all directives.

* refactor: use expansion tools in test utils