refactor(ui): replace BackButton with plain IconButton
BackButton is an unneeded component because it's exactly an IconButton from
the design system. The only thing we needed to make IconButton support
for this is hrefs, because designers to not always distinguish between
links and button, so sometimes, like in this case, what appears visually
as a button is actually a link to a different page.
We can also make Button support hrefs the same way as IconButton as some
point if we need it, it's only marginally more complicated since its
disabled prop needs to be respected too.