refactor(ui): make Anchor capable of detecting links to server-side pages
This introduces the routing package I've always wanted to build.
Right now, using `typesafe-routes` in it is kind of pointless,
but when we start using it for the "normal" client-side frontend routes,
it will be useful. I've been very happy with the package when I last
used it, its API need a bit of getting used to but then it does one job
well and it is not coupled to a specific router library.