feat: proper editor text blocks
This is a lot of lines, but it's all pretty stupid straightforward stuff 😅

If you notice heading levels being off by one in some places, this is on
purpose. It's where different systems interface. Example mapping:
`ELEMENT_H2 => Heading1 => h2`.
The design system knows `Display` as the biggest heading above Heading1.
But in HTML, we still need the proper heading 1=>2=>3=>... order.

We can probably change this to be
`ELEMENT_H1 => Heading1 => h2`.
when we refactor the document title out of the editor (where it does not
really belong, it only makes things more complicated), so that the
editor starts with `ELEMENT_H1` at what is a second-level heading both
in the design system and in HTML.

Another note:
I still expect to receive designs for outside the editor where the
designed heading styles do not match the level of the heading that we
need in HTML. Paddings might also be different from the atoms here.
We don't *need* to re-use the heading atoms in such places.
They are very light wrappers around the primitives, made because the editor
needs them as components in our current structure.
The primitives themselves are still good to use elsewhere;
there is no need to use the heading atoms everywhere;
they can stay fairly specific "document headings" if they need to;
we might even name that like that at some point.