fix crash when placing a table into another (#725)
Related to https://linear.app/decipad/issue/ENG-217/faulty-instanceof-checks, the root cause of this and probably many other issues.

When evaluating table columns, decilang needs to expand whatever you put after the '=' into a column (because tables have N rows but you can still write 'Column = 1' and have the 1 repeated). When placing a table there, the type system was expanding correctly, but not the interpreter, who mistook it for a column.