fix(editor): allow only plain text in headings
No rich text. And this also adds a whole lot of other normalization, but
styled marks in headings are the main bug that users run into that is
prevented by this.

I decided to open another PR early for this because it's getting big
again, just in terms of code size, even though it's somewhat repetitive.
There are TODO markers for the parts of the tree that I am aware are
still missing normalization.

I've added misc cleanup to existing code, such as element typings, on
the way, whenever I noticed a problem.