fix(editor): H1 duplication when pressing enter in the title
As I guessed, the problem was that the relevant nodes weren't actually
normalized at all because they were merely moved down by other nodes
appearing on top, but not changed themselves.
So it turns out editor-level normalization must actually be done to the
editor node, by traversing its children.

This looks like a complete rewrite esp. without diff
--ignore-all-space, but it's far from it 😅

The old behavior is restored. That behavior may be slightly confusing if
pressing enter at the start of the title. But even if we wanted to
change this — not in scope of this PR, which just fixes the bug
introduced by #877