feat(ui): DraggableBlock
This is only the new `ui` component, the changes needed for it,
and some related visual changes needed.
Admittedly this is kind of a stupid split because DraggableBlock is
about as tightly integrated with `editor` as it gets, and there is not
even a good way to preview in Storybook what it would realistically
look like coupled to an actual block.
But at least it should be manageable to review.