fix(editor): forbid links with empty text
It was possible to create a link with empty text in the middle of some
other text. If you then manage to get your cursor inside that link
(which was possible in some circumstances), you would start typing text
into a link that you didn't see was there.
This could mean there are hidden links in documents for years, making
the document larger and potentially being unexpectedly rediscovered by
typing something new at the right place on a years old document.

This does mean typing `[](blub)` to make a link with empty text will
result in nothing (link gets created with empty text but immediately
normalized away), but that seems very edge-casey and probably fine. We
could stop link creation and just leave the construct there but meh.