fix(auto-format-code-block): respect modifier keys
Follow up to #936 which I saw the bug in when reading the diff.
Because `is-hotkey` was replaced with a plain `.key` check and nothing
further, right now, to give a particularly absurd example,
if you press Ctrl+= (which usually means increase browser zoom level),
we catch that in an empty paragraph, convert the paragraph to a code
block, and stop the zoom change.
I'm fixing this by only handling plain `=`/`backspace` presses without
modifier keys held.