refactor(feature-flags): actually support multiple flags
This is all ~~stolen from~~ inspired by the previous open source project I was on.

It allows us to
* select which flags we want live everywhere (incl. on alpha already) by setting the flag override to true.
* select which flags we want live only in local and dev because the feature is still being worked on - this is the default.
* select which flags we want turned off completely until fixed because we noticed the feature horribly breaks everything when we shipped it by setting the flag override to false.

In tests, all flags are on, but to regression test the old behavior with the flag off to make sure everything still works for production users, `disable` can be used at the start of a test case.