feat(storybook): global padding option
Previously, for components that overflow their box, even if just with a
box shadow, we've had to add padding decorators so that all of the
component would be visible in Storybook.
This is manual and error prone.

This PR allows the storybook user to change the
padding displayed around any component to various common padding values.
This can be done using a new icon in the toolbar next to the measure,
dark mode, etc. toggles.