refactor(client,ui): abstract away toasts lib
Calls to create a toast will spread like crazy through all kinds of
components in `ui`, `editor`, and `client`, so to ever be able to switch
out this lib if we wanted, in addition to the `ToastDisplay` abstraction
we also need the library's context hook abstracted away.
Better do it now before we have a hundred usages of it spread around the code.