improve how we listen to computer signals (#2046)
* improve how we listen to computer signals

Context in React is hard to get right while having good performance. Changes in it
will re-render large chunks of the tree when we're not careful.

So we use rxjs in the computer to deal with it. However, it's not easy to fall into the pit of success when doing so.

So I've created some helpers to listen to the changes in the computer while being very efficient.

* skipping result tests for now

* fix tests

Co-authored-by: John Costa <johncosta027@gmail.com>