api(language): Deprecates spread operator (& other bits that dont matter) (#2564)
* deprecate weird table stuff

Some table stuff is just weird, and will get in the way:

 * Table spreads (creating a table out of another)
 * The select() directive (dont ask)
 * splitby()

These are "weird" because they treat tables in a specific way that we cant maintain, in a backwards-compatible way. 1. creates a separate name for the same data type, 2. is only helpful for display, the FE can use array.filter thanks, and 3. creates a table that contains other tables with different lengths.

* explain how to migrate away from ...OldTable

* make column assign not return the whole table

* fix computer tests

* properly index columns returned from table.column = ...