dx(frontend): various improvements (#6320)
* refactor(notebook-route): using same hook to simplify code

fix(depcheck)

* refactor(editor-change): storing inside notebook-state instead of context

* refactor(controller-context): migrating to notebook-state

fix(controller-hook): when controller is not defined

fix(test): adding router wrapper

* refactor(docsync-state): removing setState for docsync in top level component

* refactor(errors): removing ugly error state

* refactor(annotations): moving state to notebook-state

fix(annotations): setting permission

fix

fix

* refactor(notebook-state): using selectors everywhere to prevent unnecessary re-renders

fix

* fix(notebook-title): not updating

Helmet doesn't like the title changing it seems, and it was only working
previously because this component was re-rendered because of it's
parent.

Now that that inefficiency is gone, now we don't have that side-effect
and therefore we break the notebook title. We need to use this instead

:(

* fix(e2e): strange bug when using the same users on the same notebook

* fix(computer-columns): passing argument for row count

Fixes a very very old bug

fix(tests)

* fix(e2e): bug with error'd formula

I can't for the life of me track this bug down. I thought I had it
yesterday, and it works with `yarn serve:all` in the dev server, but not
in production builds...

* Revert "fix(computer-columns): passing argument for row count"

This reverts commit 21282e10022969a95b4551f6ce02cacb2cf40c47.

* chore(hook): making it nicer

* fix lint