fix: integrations always refreshing on multiplayer (#6924)
# Description

This is a more `misc` PR.

## Integrations Refreshing

The previous implementation of integrations refreshing wouldn't work on multiplayer because it would constantly refresh every 3 seconds or so.

Now, instead of refreshing every time `timeOfLastRun` changes, we only refresh if `timeOfLastRun` is undefined. This means that if two editors refresh the integration and get different `timeOfLastRun` values, they won't override each others and go in circles.

There is the edge case of the first time the integration is added to the notebook, which is handled by a special constant.

## Making formulas more flexibles

Formulas were pretty tighly coupled to tables, however we need to use them in integrations now, so I changed some of the hooks they rely on to be more flexible.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Release Notes

- **New Features**
  - Added a new feature flag for integration formulas.
  - Introduced a new constant to manage integration refresh behavior.

- **Bug Fixes**
  - Improved conditions for displaying UI elements in the connection table.
  - Refined integration block state management, including handling of the `timeOfLastRun` property.

- **Chores**
  - Updated feature flags configuration by removing deprecated flags.

The release enhances integration handling, feature flag management, and UI element visibility, providing improved control over application behaviors.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->